Updated docs to include all functions and classes
