+++ aotools/image_processing/__init__.py
@@ -1,3 +1,4 @@
 from .centroiders import *
 from ._image_processing import *
 from .contrast import *
+from .psf import *
