+++ doc/source/fft.rst
@@ -1,9 +1,6 @@
 Normalised Fourier Transforms
 =============================
 
-aotools.fouriertransform module
--------------------------------
-
 .. automodule:: aotools.fouriertransform
     :members:
     :undoc-members:
