+++ doc/source/pupil.rst
@@ -1,0 +1,7 @@
+Telescope Pupils
+================
+
+.. automodule:: aotools.functions.pupil
+    :members:
+    :undoc-members:
+    :show-inheritance:
