+++ doc/source/turbulence.rst
@@ -1,6 +1,19 @@
 Atmospheric Turbulence
 ======================
 
+.. automodule:: aotools.turbulence
+   :members:
+   :undoc-members:
+   :imported-members:
+   :exclude-members: CovarianceMatrix, PhaseScreenKolmogorov, PhaseScreenVonKarman, calc_slope_temporalps,
+                     calculate_structure_function, calculate_wfs_seperations, circle, cn2_to_r0, cn2_to_seeing,
+                     coherenceTime, compute_covariance_xx, compute_covariance_xy, compute_covariance_yy,
+                     create_tomographic_covariance_reconstructor, fft, fit_tps, ft_phase_screen, ft_sh_phase_screen,
+                     get_tps_time_axis, ift2, isoplanaticAngle, mirror_covariance_matrix, numpy, optimize,
+                     plot_tps, pyplot, r0_from_slopes, r0_to_cn2, r0_to_seeing, scipy, seeing_to_cn2, seeing_to_r0,
+                     slope_variance_from_r0, structure_function_kolmogorov, structure_function_vk, wfs_covariance,
+                     wfs_covariance_mpwrap, xrange
+
 .. automodule:: aotools.turbulence.phasescreen
     :members:
     :undoc-members:
